requirements.txt
Django>=4.3
djangorestframework
drf-spectacular
django-filter
django-cors-headers
psycopg2-binary
dj-database-url
gunicorn
Pillow
python-dotenv
bleach
